Feature request #25: Marked Scraper#getConnectionFactory method as deprecated.
It would be neat to decouple Scraper from ConnectionFactory. In order to achieve it, we need scoped dependency injection to work with processors (database plugin at least). Since currently we can't provide this functionality, we are not able to get rid of ConnectionFactory instance from Scraper (thus, exposing it via getter).