Feature request #24: Fixed (sometimes) failing multithreaded LockedRegistryTest.
The problem was in the test construction - all threads tried to bind and unbind value for exactly the same key. Sometimes, depending on the order of threads, 'not null' assertion failed. Of course it was valid situation - some thread might preempt the one executing assertion. Modified test in order to allow each thread to have its own, unique key and base assertion on value returned by registry for this particular key.