Feature request #31: Fixed problem with scope of variables defined within 'config' element.
Modified ConfigProcessor to retrieve list of operation definitions and for each of them create and execute processor. This logic is duplicated in BodyProcessor, however, BodyProcessor executes this operation within new context. If we had new context at the level of ConfigProcessor, variables defined within 'config' element body would not be available in the scraper context at the end of scraper's execution. This is quick fix and is eligible for refactoring as soon as we'll kick off BodyProcessor annihilation process!